+++ rcore-fs-hostfs/src/lib.rs
@@ -131,7 +131,7 @@ impl INode for HNode {
         Ok(())
     }
 
-    fn create(&self, name: &str, type_: FileType, _mode: u32) -> Result<Arc<dyn INode>> {
+    fn create(&self, name: &str, type_: FileType, _mode: u16) -> Result<Arc<dyn INode>> {
         let new_path = self.path.join(name);
         if new_path.exists() {
             return Err(FsError::EntryExist);
